// frontend/uam/src/app/gen/model/authResponse.ts
/**
 * WTM_UAM
 * User/Auth manager
 *
 * The version of the OpenAPI document: v1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AuthResponse { 
    message?: string;
    success?: boolean;
    token?: string;
    userId?: string;
}