// frontend/uam/src/app/gen/model/currentUserResponse.ts
/**
 * WTM_UAM
 * User/Auth manager
 *
 * The version of the OpenAPI document: v1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CurrentUserResponse { 
    userId?: string;
    username?: string;
}